requirements.txt
asgiref==3.11.1
boto3==1.42.78
botocore==1.42.78
dj-database-url==3.1.2
Django==6.0.3
django-cors-headers==4.9.0
django-storages==1.14.6
djangorestframework==3.17.1
gunicorn==25.3.0
jmespath==1.1.0
packaging==26.0
pillow==12.1.1
psycopg==3.3.3
psycopg-binary==3.3.3
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
s3transfer==0.16.0
six==1.17.0
sqlparse==0.5.5
tzdata==2025.3
urllib3==2.6.3
whitenoise==6.12.0
